Add files via upload
•	In this problem we are not provided with target variable explicitly. 
•	Last working Date is used to find the target variable. 
•	If Last working date is given then that person would have resigned in that month. if the person decided to leave then he would have decided earlier itself and we were asked to predict if the person will resign in next 2 quarters. So we have chosen target based on that.

The scoring criteria is F1 score. We have to predict the persons leaving correctly. It is generally easy to get good accuracy in such problems but not F1 score .

We have created new features like experience days , salary increased, no of promotions, rating increased or decreased to get the final results , We got an F1 score of 79 % using Random Forest Classifier